Move dump_session(), load_session() and everything related to session.py
